Refactor complete and tested
