+++ src/python/eight_bit_computer/operations/set_op.py
@@ -12,7 +12,7 @@ from ..language_defs import (
     FLAGS,
     instruction_byte_from_bitdefs
 )
-from operation_utils import assemble_instruction, match_and_parse_line
+from ..operation_utils import assemble_instruction, match_and_parse_line
 from ..data_structures import (
     get_arg_def_template, get_machine_code_byte_template
 )
