+++ src/python/eight_bit_computer/token_utils.py
@@ -1,6 +1,7 @@
 """
 
 """
+import re
 
 
 def is_label(test_string):
