+++ src/python/tests/test_assembly_validity.py
@@ -38,11 +38,11 @@ def test_check_multiple_label_defs_raises():
     lines.append(line)
 
     with pytest.raises(AssemblyError):
-        validity.check_multiple_label_defs(lines)
+        assembly_validity.check_multiple_label_defs(lines)
 
 
 def test_check_multiple_label_defs_does_not_raise(processed_assembly_lines):
-    validity.check_multiple_label_defs(processed_assembly_lines)
+    assembly_validity.check_multiple_label_defs(processed_assembly_lines)
     assert True
 
 
@@ -93,11 +93,11 @@ def test_check_multiple_label_assignment_raises():
     lines.append(line)
 
     with pytest.raises(AssemblyError):
-        validity.check_multiple_label_assignment(lines)
+        assembly_validity.check_multiple_label_assignment(lines)
 
 
 def test_check_multiple_label_assignment_does_not_raise(processed_assembly_lines):
-    validity.check_multiple_label_assignment(processed_assembly_lines)
+    assembly_validity.check_multiple_label_assignment(processed_assembly_lines)
     assert True
 
 
@@ -139,11 +139,11 @@ def test_check_undefined_label_ref_raises():
     lines.append(line)
 
     with pytest.raises(AssemblyError):
-        validity.check_undefined_label_ref(lines)
+        assembly_validity.check_undefined_label_ref(lines)
 
 
 def test_check_undefined_label_ref_does_not_raise(processed_assembly_lines):
-    validity.check_undefined_label_ref(processed_assembly_lines)
+    assembly_validity.check_undefined_label_ref(processed_assembly_lines)
     assert True
 
 
@@ -178,11 +178,11 @@ def test_check_multiple_variable_def_raises():
     lines.append(line)
 
     with pytest.raises(AssemblyError):
-        validity.check_multiple_variable_def(lines)
+        assembly_validity.check_multiple_variable_def(lines)
 
 
 def test_check_multiple_variable_def_does_not_raise(processed_assembly_lines):
-    validity.check_multiple_variable_def(processed_assembly_lines)
+    assembly_validity.check_multiple_variable_def(processed_assembly_lines)
     assert True
 
 
@@ -199,7 +199,7 @@ def test_check_num_variables_raises():
         lines.append(line)
 
     with pytest.raises(AssemblyError):
-        validity.check_num_variables(lines, 0)
+        assembly_validity.check_num_variables(lines, 0)
 
 
 def test_check_num_variables_offset_raises():
@@ -215,11 +215,11 @@ def test_check_num_variables_offset_raises():
         lines.append(line)
 
     with pytest.raises(AssemblyError):
-        validity.check_num_variables(lines, 250)
+        assembly_validity.check_num_variables(lines, 250)
 
 
 def test_check_num_variables_does_not_raise(processed_assembly_lines):
-    validity.check_num_variables(processed_assembly_lines, 0)
+    assembly_validity.check_num_variables(processed_assembly_lines, 0)
     assert True
 
 
@@ -254,9 +254,9 @@ def test_check_num_instruction_bytes_raises():
         lines.append(line)
 
     with pytest.raises(AssemblyError):
-        validity.check_num_instruction_bytes(lines)
+        assembly_validity.check_num_instruction_bytes(lines)
 
 
 def test_check_num_instruction_bytes_does_not_raise(processed_assembly_lines):
-    validity.check_num_instruction_bytes(processed_assembly_lines)
+    assembly_validity.check_num_instruction_bytes(processed_assembly_lines)
     assert True
